requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.2
psnawp_api==2.1.0
pytest==8.3.2
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
selenium==4.23.1
webdriver_manager==4.0.2